Create SecuritySimulearn.md
The certificate was acquired from the AWS skillsbuilder platform specifically focusing on the core security concepts of the cloud. The key takeaways included an in depth understanding of Amazon EC2 and IAM users, roles, and groups and their creation processes. I got an understanding of the AWS Shared Responsibility Model and compliance programs, as well.